Moved ExportResult from MainMenu to Experimenter
