+++ src/Model/FrontEndAPI/FrontEndExperimenter.cs
@@ -1,6 +1,7 @@
+namespace Model.FrontEndAPI;
 using Model.Result;
 using Model.Database;
-namespace Model.FrontEndAPI;
+using System.Text;
 internal class FrontEndExperimenter : IFrontEndExperimenter {
 
     private IDatabase databaseService;
@@ -11,7 +12,25 @@ internal class FrontEndExperimenter : IFrontEndExperimenter {
     public void StoreResultFromQuestion(IResult answer) {
         databaseService.StoreResult(answer);   
     }
+    
     public int GetNextUserId() {
         return databaseService.GetNextUserId();
     }
+    
+    public bool ExportResults(int surveyWrapperId, string folderPath) {
+        List<Result> results = databaseService.GetSurveyWrapperResults(surveyWrapperId);
+        string path = Path.Combine(folderPath, $"{surveyWrapperId}.csv");
+        try {
+            using (StreamWriter writer = new StreamWriter(path, false, Encoding.UTF8)) {
+                foreach (var result in results) {
+                    writer.WriteLine(result.ToString());
+                }
+            }
+            return true;
+        }
+        catch (Exception) {
+            return false;
+        } 
+    }
+
 }
