+++ src/Model/FrontEndAPI/FrontEndMainMenu.cs
@@ -1,9 +1,7 @@
 namespace Model.FrontEndAPI;
 using Model.Database;
-using Model.Result;
 using Model.Survey;
-using System.Text;
-using System.IO;
+
 
 internal class FrontEndMainMenu : IFrontEndMainMenu {
 
@@ -13,21 +11,7 @@ internal class FrontEndMainMenu : IFrontEndMainMenu {
         db = database;
     }
 
-    public bool ExportResults(int surveyWrapperId, string folderPath) {
-        List<Result> results = db.GetSurveyWrapperResults(surveyWrapperId);
-        string path = Path.Combine(folderPath, $"{surveyWrapperId}.csv");
-        try {
-            using (StreamWriter writer = new StreamWriter(path, false, Encoding.UTF8)) {
-                foreach (var result in results) {
-                    writer.WriteLine(result.ToString());
-                }
-            }
-            return true;
-        }
-        catch (Exception) {
-            return false;
-        } 
-    }
+
     public IReadOnlySurveyWrapper? GetSurveyWrapper(int surveyId) {
         return db.GetSurveyWrapper(surveyId);
     }
