+++ src/Model/FrontEndAPI/IFrontEndExperimenter.cs
@@ -5,4 +5,5 @@ public interface IFrontEndExperimenter {
     
     void StoreResultFromQuestion(IResult answer);
     int GetNextUserId();
+    bool ExportResults(int surveyWrapperId, string folderPath);
 }
