+++ src/Model/FrontEndAPI/IFrontEndMainMenu.cs
@@ -5,7 +5,7 @@ public interface IFrontEndMainMenu {
     List<IModifySurveyWrapper>? ValidateSuperUser(string username, string password);
     bool ImportSurveyWrapper(string filePath);
 
-    bool ExportResults(int surveyWrapperId, string folderPath);
+
     IReadOnlySurveyWrapper? GetSurveyWrapper(int pincode);
 
 }
